requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Task Queue
celery==5.3.4
redis==5.0.1
kombu==5.3.4

# HTTP & Scraping
httpx==0.25.1
requests==2.31.0
trafilatura==1.6.3
readability-lxml==0.8.1
urllib3==2.1.0
beautifulsoup4==4.12.2
lxml==4.9.3

# PDF Generation
weasyprint==60.1
pdfkit==1.0.0

# Data Processing
pydantic==2.5.0
pydantic-settings==2.1.0
python-dateutil==2.8.2
pytz==2023.3
simhash==2.1.2

# LLM (Optional)
openai==1.3.5

# Email
aiosmtplib==3.0.1

# Utilities
pyyaml==6.0.1
click==8.1.7

# Language Detection (for India multilingual support)
langid==1.1.6
langdetect==1.0.9

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.1
responses==0.24.1

# Code Quality
ruff==0.1.6
mypy==1.7.1
types-requests==2.31.0.10
types-python-dateutil==2.8.19.14

